refactor(types): accordion
