+++ src/lib/builders/accordion/types.ts
@@ -1,0 +1,19 @@
+import type { createAccordion } from './create';
+
+type BaseAccordionArgs = {
+	disabled?: boolean;
+};
+
+type SingleAccordionArgs = {
+	value?: string;
+	type?: 'single';
+};
+
+type MultipleAccordionArgs = {
+	value?: string[];
+	type: 'multiple';
+};
+
+export type CreateAccordionArgs = BaseAccordionArgs & (SingleAccordionArgs | MultipleAccordionArgs);
+
+export type CreateAccordionReturn = ReturnType<typeof createAccordion>;
